app/requirements.txt
flask==3.0.0
mysql-connector-python==8.3.0
pytest==7.4.0
